Add Pieces to Concerts
Create new metadata in the concert custom post type called "pieces".
This is a JSON string which is held in a single field in the metadata
table. Javascript in the admin pages picks this up and renders it into
the new "concert-pieces" metabox. There are a couple of outstanding
actions:

- The javascript loads on every admin page. This is messy.
- The order of pieces in the concert cannot yet be edited.