Load javascript and css only on appropriate post page
Previously the javascript and css for an abstract meta box was loaded on
every admin page. This has now been fixed, so that the scripts and
styles are only loaded on the post.php, and post-new.php page for the
appropriate post type. The AbstractMetaBox constructor now takes an
additional paramater which is the post type to which the MetaBox
pertains. Therefore there is now an additional abstract class -
AbstractConcertMetaBox, which sets this to 'concert'.