Removed MetaBox nonce logic
There appears to be some contension as to whether or not a nonce is
required. I've also corrected the save hook which was being used - I was
on save_post not save_post_{post_type} so the save method of
abstract_meta_box was being called nearly all the time causing random
data issues.